pyproject.toml
[project]
name = "stochastic-credit-risk-and-ecl-engine"
version = "0.1.0"
description = ""
authors = [
    {name = "amandipd",email = "amandip@berkeley.edu"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "numpy (>=2.4.1,<3.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
